Just POCing what a global context would do
Right now we have a lot of constructors that take in the context, which is really hard to use outside of the behave "context". 

However, it would be nice to be able to have a classmethod without a context arg, that attempts to get the context and unpack the values into reasonable attributes. 

We should also consider a CustomConfig object that we use _instead_ of the context. We could then pass the CustomConfig around in the context for use in behave steps.